requirements.txt
alabaster==0.7.12
Babel==2.10.3
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
dis==1.0.1
docutils==0.19
greenlet==1.1.2
idna==3.3
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pycryptodomex==3.15.0
Pygments==2.12.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.37
tabulate==0.8.9
urllib3==1.26.11